Add a basic PathCommandProvider
This checks for the given command name in each directory in $PATH, and
returns an object for executing the first match it finds.

In situations where we don't need to redirect output, we use node-pty to
spawn a new PTY for the command to run in. This works better than our
own ioctl implementation for now, but but can't be used always because
it doesn't allow distinguishing between stderr and stdout. So when we
do need to redirect those, we use Node's process spawning.

Current issues:

- The unhandledRejection callback is very dubious.
- We eat one chunk of stdin input after the executable stops, because we
  can't cancel `ctx.externs.in_.read()`. Possibly this should go via
  another stream that we can disconnect.
- Stdin is always echoed even when the command handles it for us. This
  means typing in the `python` repl makes each character appear
  doubled. It's not actually doubled though, it's just a visual error.