Stop including PathCommandProvider (and so node-pty) in web build
This is only usable in a native terminal. Including it also makes rollup
upset about having to include a binary file from the node-pty library.

There is a small hack here: rollup will automatically include anything
that's imported, but splitting the imported path into two strings
concatenated is enough to fool it into ignoring the file.